--- src/main/java/com/nexters/goalpanzi/application/member/dto/ProfileRequest.java
@@ -1,13 +1,0 @@
-package com.nexters.goalpanzi.application.member.dto;
-
-import com.nexters.goalpanzi.domain.member.CharacterType;
-import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.NotNull;
-
-public record ProfileRequest(
-        @Schema(description = "닉네임", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-        String nickname,
-        @Schema(description = "캐릭터 타입", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-        @NotNull CharacterType characterType
-) {
-}
