+++ src/main/java/com/nexters/goalpanzi/presentation/mission/MissionControllerDocs.java
@@ -1,8 +1,8 @@
 package com.nexters.goalpanzi.presentation.mission;
 
-import com.nexters.goalpanzi.application.mission.dto.CreateMissionRequest;
-import com.nexters.goalpanzi.application.mission.dto.MissionResponse;
+import com.nexters.goalpanzi.application.mission.dto.MissionDto;
 import com.nexters.goalpanzi.common.argumentresolver.LoginMemberId;
+import com.nexters.goalpanzi.presentation.mission.dto.CreateMissionRequest;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 @Tag(name = "미션")
 public interface MissionControllerDocs {
 
-    ResponseEntity<MissionResponse> createMission(
+    ResponseEntity<MissionDto.MissionResponse> createMission(
             @Parameter(hidden = true) @LoginMemberId final Long memberId,
             @Valid @RequestBody final CreateMissionRequest request
     );
